requirements: Use fixed versions
When deploying onto a new server things broke again due to
incompatibilies in SQLAlchemy. This fixes it by using the exact packet
versions used on the old server so they date back to the date of the
previous commit. The file has been generated by running "./venv/pip
freeze -l" on the old server.

Signed-off-by: Konrad Gräfe <kgraefe@paktolos.net>